+++ space_battle/src/server/game_elements/HalfScores.java
@@ -4,7 +4,7 @@ public class HalfScores extends PowerDown{
 	
 	private static int timeItLasts = 15000; // 15 sec
 	
-	public HalfScores(int x, int y){
+	public HalfScores(double x, double y){
 		super(x, y);
 	}
