+++ space_battle/src/server/game_elements/LeftRightSwitcher.java
@@ -4,7 +4,7 @@ public class LeftRightSwitcher extends PowerDown{
 	
 	private static int timeItLasts = 20000; // 20 sec
 	
-	public LeftRightSwitcher(int x, int y){
+	public LeftRightSwitcher(double x, double y){
 		super(x, y);
 	}
