+++ space_battle/src/server/game_elements/NoAmmo.java
@@ -4,7 +4,7 @@ public class NoAmmo extends PowerDown{
 	
 	private static int timeItLasts = 10000; // 10 sec
 	
-	public NoAmmo(int x, int y){
+	public NoAmmo(double x, double y){
 		super(x, y);
 	}
